refactor: make util tests import from top level to get 100 coverage
