+++ src/util/__test__/compose-test.js
@@ -1,6 +1,6 @@
 // @flow
 
-import compose from '../compose';
+import compose from '../../compose';
 
 test('compose composes from right to left', () => {
     const double = x => x * 2;
